fix(front): titleHeadline renamed to decoratedHeadline
