+++ redi-connect-front/src/components/atoms/DecoratedHeadline.scss
@@ -1,6 +1,6 @@
 @import "~bulma/sass/utilities/_all";
 
-.title-headline {
+.decorated-headline {
   padding-bottom: 3.5rem;
 
   @include mobile {
